refactor(#188): move controllers to a new files
